refactor : 사용자 관련 서비스 및 jpa 구현체 수정
