--- src/main/java/roomescape/domain/member/domain/repository/MemberRepositoryCustom.java
@@ -1,6 +1,0 @@
-package roomescape.domain.member.domain.repository;
-
-public interface MemberRepositoryCustom {
-
-    Long updateAdminRole(Long id);
-}
